Moved Domain layer to SPM
