+++ FinderLost/Library/Data/Source/Firebase/GoogleAuthProvider.swift
@@ -5,8 +5,12 @@
 //  Created by Andres Felipe Alzate Restrepo on 8/2/23.
 //
 
+import Domain
+
 import Combine
+
 import Factory
+
 import Firebase
 import GoogleSignIn
