+++ Packages/Domain/Sources/Domain/Entity/CustomError.swift
@@ -5,6 +5,6 @@
 //  Created by Andres Felipe Alzate Restrepo on 8/2/23.
 //
 
-enum CustomError: Error {
+public enum CustomError: Error {
     case isEmpty
 }
