+++ Packages/Domain/Sources/Domain/Entity/UserSession.swift
@@ -5,12 +5,7 @@
 //  Created by Andres Felipe Alzate Restrepo on 8/2/23.
 //
 
-protocol UserSession {
+public protocol UserSession {
     var token: String { get }
     var userId: String { get }
-}
-
-struct UserSessionImpl: UserSession {
-    let token: String
-    let userId: String
 }
