+++ Packages/Domain/Sources/Domain/Repository/UserRepository.swift
@@ -1,0 +1,14 @@
+//
+//  UserRepository.swift
+//  
+//
+//  Created by Andres Felipe Alzate Restrepo on 11/2/23.
+//
+
+import Combine
+
+public protocol UserRepository {
+    func getSession() -> AnyPublisher<UserSession, Error>
+    func signIn() -> AnyPublisher<UserSession, Error>
+    func signOut() -> AnyPublisher<Void, Error>
+}
